Renaming Storage base classes, Getting rid of unnecessary aimrocks dependencies
